Move DropGem from components to pages
+ move out shared components out of shared folder
