+++ frontend/src/components/FriendCell.vue
