+++ frontend/src/components/GemHeader.vue
