+++ frontend/src/pages/DropGem/GemColorCarousel.vue
